+++ tests/Containers/Interfaces/Factories/IStackInventoryFactory.cs
@@ -7,5 +7,6 @@ namespace TheChest.Inventories.Tests.Containers.Interfaces.Factories
         IStackInventory<T> EmptyContainer(int size = 20, int stackSize = 10);
         IStackInventory<T> FullContainer(int size, int stackSize, T item = default!);
         IStackInventory<T> ShuffledItemsContainer(int size, int stackSize, params T[] items);
+        IStackInventory<T> WithItems(T[] items, int stackSize, int size);
     }
 }